View Has been created
