+++ src/View/CadLivros.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package br.com.SGAL.Telas;
+package View;
 
 import java.sql.*;
 import br.com.SGAL.dal.ModelConection;
